+++ config/desktop/firefox.desktop
@@ -1,0 +1,10 @@
+[Desktop Entry]
+
+Name=Firefox
+GenericName=Firefox
+Exec=/opt/firefox/firefox
+Terminal=false
+Icon=/opt/firefox/browser/chrome/icons/default/default128.png
+Type=Application
+Categories=Application;Network;X-Developer;
+Comment=Firefox
